requirements.txt
fastapi
uvicorn
langchain
google-generativeai
langchain-google-genai
pandas
python-dotenv
streamlit
requests
